+++ example/app/forms.py
@@ -47,7 +47,9 @@ class TestForm(forms.Form):
     category5 = forms.ChoiceField(widget=RadioSelectButtonGroup, choices=MEDIA_CHOICES)
     color = forms.CharField(widget=TextInput(attrs={"type": "color"}))
     range = forms.IntegerField(widget=TextInput(attrs={"type": "range"}))
-    addon = forms.CharField(widget=forms.TextInput(attrs={"addon_before": "before", "addon_after": "after"}))
+    addon = forms.CharField(
+        widget=forms.TextInput(attrs={"addon_before": "before", "addon_after": "after"}), help_text="Addons work!"
+    )
     date5 = forms.DateField(widget=TextInput(attrs={"type": "date"}))
     time5 = forms.TimeField(widget=TextInput(attrs={"type": "time"}))
     url5 = forms.CharField(widget=TextInput(attrs={"type": "url"}))